# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
  internal_rest_api_artist.GetArtistByIDOut:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  internal_rest_api_artist.PatchUpdateArtistIn:
    properties:
      artistID:
        type: integer
      bio:
        type: string
      name:
        type: string
    required:
    - artistID
    type: object
  internal_rest_api_artist.PatchUpdateArtistOut:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  internal_rest_api_artist.PatchUpdateAvatarOut:
    properties:
      avatar_url:
        type: string
    type: object
  internal_rest_api_artist.PostArtistIn:
    properties:
      bio:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  internal_rest_api_artist.PostArtistOut:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: lyryx API
  version: "1.0"
paths:
  /v1/artist:
    post:
      consumes:
      - application/json
      description: Создает нового артиста по имени и био. Доступно только модератору.
      parameters:
      - description: Параметры артиста
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_rest_api_artist.PostArtistIn'
      produces:
      - application/json
      responses:
        "201":
          description: Артист успешно создан
          schema:
            $ref: '#/definitions/internal_rest_api_artist.PostArtistOut'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Пользователь не аутентифицирован
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Имя артиста уже занято
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Создание артиста.
      tags:
      - artist
  /v1/artist/{id}:
    get:
      description: Возвращает полную информацию о профиле артиста по его id.
      parameters:
      - description: Artist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с профилем артиста
          schema:
            $ref: '#/definitions/internal_rest_api_artist.GetArtistByIDOut'
        "400":
          description: Некорректный id артиста
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Артист не найден
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Получение данных артиста по его id.
      tags:
      - artist
    patch:
      consumes:
      - application/json
      description: Обновляет name, bio артиста по его id. Достаточно передать только
        нужные поля.
      parameters:
      - description: Параметры обновления профиля артиста
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_rest_api_artist.PatchUpdateArtistIn'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль артиста обновлен
          schema:
            $ref: '#/definitions/internal_rest_api_artist.PatchUpdateArtistOut'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Пользователь не аутентифицирован
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Артист не найден
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Артист с таким именем уже существует
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Частичное обновление профиля артиста
      tags:
      - artist
  /v1/artist/{id}/avatar:
    patch:
      consumes:
      - multipart/form-data
      description: Принимает avatar в multipart/form-data, валидирует изображение,
        загружает в MinIO и обновляет ссылку на аватар артиста.
      parameters:
      - description: Файл аватарки (png/jpeg)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Аватар артиста обновлен
          schema:
            $ref: '#/definitions/internal_rest_api_artist.PatchUpdateAvatarOut'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Пользователь не аутентифицирован
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Артист не найден
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Обновление аватарки артиста
      tags:
      - artist
...
swagger: "2.0"